Picker CreationForm: warn on missing real id, keep form
Log a warning when the save "succeeded" but the tempid has not been mapped to a real one, indicating that it actually failed.

The form will be kept open and reset (the latter is not optimal...).

Notice that if the backend properly returns ::form/errors then we won't hit this code path.